fix: unify LSP server PATH resolution between detection and spawn (#2431)
