+++ src/tools/lsp/server-path-bases.ts
@@ -1,0 +1,15 @@
+import { join } from "path"
+
+import { getDataDir, getOpenCodeConfigDir } from "../../shared"
+
+export function getLspServerAdditionalPathBases(workingDirectory: string): string[] {
+  const configDir = getOpenCodeConfigDir({ binary: "opencode" })
+  const dataDir = join(getDataDir(), "opencode")
+
+  return [
+    join(workingDirectory, "node_modules", ".bin"),
+    join(configDir, "bin"),
+    join(configDir, "node_modules", ".bin"),
+    join(dataDir, "bin"),
+  ]
+}
